src/vehicle-manager/pyproject.toml
[project]
name = "vehicle-manager"
version = "0.1.0"
requires-python = "==3.14.*"
dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "fastapi-problem>=0.12.0",
    "fastapi[standard]>=0.128.0",
    "nats-py>=2.12.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pyjwt[crypto]>=2.10.1",
    "pytest>=9.0.2",
    "shapely>=2.1.2",
    "sqlalchemy>=2.0.45",
]

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.pyright]
typeCheckingMode = "standard"

[tool.pytest.ini_options]
testpaths = ["src/vehicle_manager/tests/"]
